cleanup codegen structure
