+++ mlpfile/__init__.py
@@ -3,13 +3,15 @@
 import tempfile
 import ctypes
 import subprocess
+import warnings
 
 import numpy as np
 
 from _mlpfile import Model, Layer, LayerType
+from ._codegen import codegen_c, codegen_eigen
 
 
-__all__ = ["Model", "Layer", "LayerType", "codegen_compile", "codegen_c", "codegen_eigen", "cpp_dir"]
+__all__ = ["Model", "Layer", "LayerType", "codegen", "cpp_dir"]
 
 
 def cpp_dir():
@@ -24,20 +26,20 @@ def cpp_dir():
     print(os.path.dirname(__file__) + "/cpp", end="")  # no newline
 
 
-def codegen_compile(model, workdir=None, eigen=True):
-    """Generates and compiles code to evaluate the model's forward pass.
+def codegen(model, outdir=None, eigen=True, compile=False):
+    """Generates and optionally compiles C/C++ code to evaluate the model's
+    forward pass.
 
-    Compiles the code into an object file and shared library containing the
-    function::
+    The generated code has the function::
 
         void forward(float const *x, float *y)
 
     Note that size arguments are not needed because they are known at compile time.
 
-    This function also loads the library with ``ctypes`` and returns the Python
-    object representing the library. To call from Python+NumPy, use code like::
+    If ``compile=True``, also compiles the code into a shared library and loads
+    that library with ``ctypes``. Calling from Python+NumPy looks like::
 
-        lib = mlpfile.codegen_compile(model)
+        lib = mlpfile.codegen(model, compile=True)
         x = <my input of length model.input_dim()>.astype(np.float32)
         y = np.zeros(model.output_dim(), dtype=np.float32)
         xptr = x.ctypes.data_as(ctypes.POINTER(ctypes.c_float))
@@ -45,17 +47,20 @@ def codegen_compile(model, workdir=None, eigen=True):
         lib.forward(xptr, yptr)
         # Now y has the output.
 
-    However, even if you want to use the code from C instead of Python, this
-    function can still be helpful because it will use the correct compiler
-    flags. This is especially important when ``eigen=True``.
+    Even if you don't want to use the compiled code from Python, this function
+    can be helpful because it knows the correct compiler flags. This is
+    especially important when ``eigen=True``.
+
+    To compile the Eigen code yourself, these flags ensure the code works in a
+    malloc-free (embedded/realtime) environment and can be linked as C::
+
+        -fno-exceptions -DEIGEN_NO_MALLOC
 
     Args:
         model (Model): Model.
-        workdir (str, optional): If specified, writes code, object file, and
+        outdir (str, optional): If specified, writes code, object file, and
             shared lib to this directory. Otherwise, uses a temporary directory
-            deleted immediately after call. The returned ``ctypes.CDLL`` will be
-            loaded before the temporary dir is deleted, so you can still use it
-            until the process terminates.
+            deleted immediately after call. ``outdir=None`` only makes sense when ``compile=True``.
         eigen (bool): Use Eigen internally. If true, we generate code that uses
             the Eigen C++ library. Eigen's matrix-vector multiply is much
             faster than our hand-written naive one when ``eigen=False``.
@@ -63,24 +68,40 @@ def codegen_compile(model, workdir=None, eigen=True):
             memory, and no exceptions. Therefore, the compiled object/library
             can be used in C projects, with a C linker, even though it was
             compiled by a C++ compiler.
+        compile (bool): Compiles the code into a shared library and loads it
+            with ``ctypes``. If ``outdir=None``, the returned ``ctypes.CDLL``
+            will be loaded before the temporary dir is deleted, so you can
+            still use the library for the remainder of the Python process.
 
     Returns:
-        a ``ctypes.CDLL`` library with function ``forward`` described in overview.
+        If ``compile=True``, returns a ``ctypes.CDLL`` library with function
+        ``forward`` described in overview. Otherwise, returns nothing.
     """
-    if workdir is None:
-        with tempfile.TemporaryDirectory() as d:
-            return codegen_compile(model, d, eigen)
+    if not compile and outdir is None:
+        warnings.warn(
+            "Will have no observable effect (besides errors)"
+            " with compile=False and outdir=None."
+        )
 
-    src = workdir + "/src.c" + ("pp" if eigen else "")
-    obj = workdir + "/obj.o"
-    lib = workdir + "/lib.so"
+    if outdir is None:
+        with tempfile.TemporaryDirectory() as d:
+            return codegen(model, outdir=d, eigen=eigen, compile=compile)
+
+    src = outdir + "/src." + ("cpp" if eigen else "c")
+    obj = outdir + "/obj.o"
+    lib = outdir + "/lib.so"
     with open(src, "w") as f:
         if eigen:
             codegen_eigen(model, f)
         else:
             codegen_c(model, f)
+
+    if not compile:
+        return
+
     flags_both = ["-c", "-fPIC", "-O3", "-o", obj]
     if eigen:
+        # TODO: Useful to allow getting Eigen somewhere else besides eigenpip?
         import eigenpip
         incl = "-I" + eigenpip.get_include()
         flags_eig = [incl, "-fno-exceptions", "-DEIGEN_NO_MALLOC"]
@@ -91,136 +112,3 @@ def codegen_compile(model, workdir=None, eigen=True):
     result_shared = subprocess.run(["cc", "-shared", "-o", lib, obj])
     assert result_shared.returncode == 0
     return ctypes.cdll.LoadLibrary(lib)
-
-
-def codegen_c(model, f):
-    """Generates C code to evaluate the model's forward pass.
-
-    Takes ``mlpfile.Model`` and open file-like object. See the docstring of
-    ``codegen_compile`` for details.
-    """
-    def array1d(x):
-        return "{" + ", ".join([str(xi) for xi in x]) + "}"
-
-    def array2d(X):
-        return "{\n" + ",\n".join([array1d(row) for row in X]) + "\n}"
-
-    # static data
-    for i, layer in enumerate(model.layers):
-        if layer.type == LayerType.Linear:
-            rows, cols = layer.W.shape
-            f.write(f"float W_{i}[{rows}][{cols}] = {array2d(layer.W)};\n")
-            f.write(f"float b_{i}[{rows}] = {array1d(layer.b)};\n")
-
-    # workspace
-    linears = [layer for layer in model.layers if layer.type == LayerType.Linear]
-    workdim = max(
-        [lay.W.shape[0] for lay in linears]
-        + [lay.W.shape[1] for lay in linears]
-    )
-    f.write(f"float work[2][{workdim}];\n")
-
-    # code
-    size = model.input_dim()
-    f.write(f"void forward(float const *x, float *y) {{")
-    f.write(
-f"""
-for (int i = 0; i < {size}; ++i) {{
-    work[0][i] = x[i];
-}}
-""")
-    src = 0
-    dst = 1
-    for ilayer, layer in enumerate(model.layers):
-        if layer.type == LayerType.Input:
-            continue
-        elif layer.type == LayerType.Linear:
-            newsize = layer.W.shape[0]
-            f.write(
-f"""
-for (int i = 0; i < {newsize}; ++i) {{
-    work[{dst}][i] = b_{ilayer}[i];
-    for (int j = 0; j < {size}; ++j) {{
-        work[{dst}][i] += W_{ilayer}[i][j] * work[{src}][j];
-    }}
-}}
-""")
-            size = newsize
-        elif layer.type == LayerType.ReLU:
-            # Avoid importing math.h just for fmaxf.
-            f.write(
-f"""
-for (int i = 0; i < {size}; i++) {{
-    work[{dst}][i] = (work[{src}][i] < 0.0f) ? 0.0f : work[{src}][i];
-}}
-""")
-        else:
-            raise ValueError("layer type:", layer.type)
-        src, dst = dst, src
-    f.write(
-f"""
-for (int i = 0; i < {size}; ++i) {{
-    y[i] = work[{src}][i];
-}}
-}}
-""")
-
-
-def codegen_eigen(model, f):
-    """Generates C++ code with C-only API to evaluate the model's forward pass.
-
-    Takes ``mlpfile.Model`` and open file-like object. See the docstring of
-    ``codegen_compile`` for details.
-    """
-    def array1d(x):
-        return "{" + ", ".join([str(xi) for xi in x]) + "}"
-
-    def array2d(X):
-        return "{\n" + ",\n".join([array1d(row) for row in X]) + "\n}"
-
-    f.write("#include <Eigen/Dense>\n")
-    f.write('static_assert(EIGEN_HAS_CXX11, "we need C++11");\n')
-
-    # static data
-    for i, layer in enumerate(model.layers):
-        if layer.type == LayerType.Linear:
-            rows, cols = layer.W.shape
-            # I would have preferred to store data directly in Eigen::Matrices
-            # instead of the extra complexity of these Eigen::Maps. But for
-            # some reason the initializer list constructor of Eigen::Matrix
-            # wasn't working even though I was using C++11.
-            f.write(f"float arr_W_{i}[{rows}][{cols}] = {array2d(layer.W)};\n")
-            f.write(f"float arr_b_{i}[{rows}] = {array1d(layer.b)};\n")
-            f.write(f"Eigen::Map<Eigen::Matrix<float, {rows}, {cols}, Eigen::RowMajor>> W_{i} (arr_W_{i}[0]);\n")
-            f.write(f"Eigen::Map<Eigen::Matrix<float, {rows}, 1>> b_{i} (arr_b_{i});\n")
-
-    # workspace
-    linears = [layer for layer in model.layers if layer.type == LayerType.Linear]
-    workdim = max(
-        [lay.W.shape[0] for lay in linears]
-        + [lay.W.shape[1] for lay in linears]
-    )
-    f.write(f"Eigen::Matrix<float, {workdim}, 1> work[2];\n")
-
-    # code
-    size = model.input_dim()
-    f.write(f'extern "C" void forward(float const *xptr, float *yptr) {{\n')
-    f.write(f"Eigen::Map<Eigen::Matrix<float, {size}, 1> const> x(xptr);\n")
-    f.write(f"work[0].template head<{size}>() = x;\n")
-    src = 0
-    dst = 1
-    for ilayer, layer in enumerate(model.layers):
-        if layer.type == LayerType.Input:
-            continue
-        elif layer.type == LayerType.Linear:
-            newsize = layer.W.shape[0]
-            f.write(f"work[{dst}].template head<{newsize}>() = W_{ilayer} * work[{src}].template head<{size}>() + b_{ilayer};\n")
-            size = newsize
-        elif layer.type == LayerType.ReLU:
-            f.write(f"work[{dst}].template head<{size}>() = work[{src}].template head<{size}>().array().max(0);\n")
-        else:
-            raise ValueError("layer type:", layer.type)
-        src, dst = dst, src
-    f.write(f"Eigen::Map<Eigen::Matrix<float, {model.output_dim()}, 1>> y(yptr);\n")
-    f.write(f"y = work[{src}].template head<{size}>();\n")
-    f.write("}")
